requirements.txt
# AI Doctor Pro — Full Stack Dependencies

# ── FastAPI backend ─────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9        # file uploads
pydantic[email]>=2.7.0

# ── Auth ────────────────────────────────────────────────────────────────────────
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0

# ── Database ────────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.0

# ── AI / API clients ────────────────────────────────────────────────────────────
groq>=0.15.0
elevenlabs>=1.50.3

# ── RAG ─────────────────────────────────────────────────────────────────────────
chromadb>=0.5.0
sentence-transformers>=3.0.0

# ── Audio ───────────────────────────────────────────────────────────────────────
gtts>=2.5.4
speechrecognition>=3.13.0
pydub>=0.25.1

# ── Frontend ────────────────────────────────────────────────────────────────────
gradio>=5.12.0
requests>=2.32.3

# ── Utilities ───────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0